3.2. Конвертер валюты

+++ chapter_001/src/test/java/ru/job4j/converter/ConverterTest.java
@@ -10,26 +10,26 @@ public class ConverterTest {
     @Test
     public void when60RubleToDollarThen1() {
         Converter converter = new Converter();
-        int result = converter.rubleToDollar(60);
+        double result = converter.rubleToDollar(60);
         assertThat(result, is(1));
     }
 
     @Test
     public void when70RubleToEuroThen1() {
         Converter converter = new Converter();
-        int result = converter.rubleToEuro(70);
+        double result = converter.rubleToEuro(70);
         assertThat(result, is(1));
     }
     @Test
     public void when3EuroToRubleThen210() {
         Converter converter = new Converter();
-        int result = converter.euroToRuble(3);
+        double result = converter.euroToRuble(3);
         assertThat(result, is(210));
     }
     @Test
     public void when2DollarToRubleThen120() {
         Converter converter = new Converter();
-        int result = converter.dollarToRuble(2);
+        double result = converter.dollarToRuble(2);
         assertThat(result, is(120));
     }
 }
